// src/utils/enum.ts
export enum ScreenName {
  Home = 'HomeScreen',
  ChangeDeliveryAddress = 'ChangeDeliveryAddressScreen',
  ChooseAddressOnMapScreen = 'ChooseAddressOnMapScreen',
  AskLocation = 'AskLocationScreen',
  Category = 'CategoryScreen',
  Note = 'NoteScreen',
  Coupon = 'CouponScreen',
  Review = 'ReviewScreen',
  PaymentFood = 'PaymentFoodScreen',
  FoodCart = 'FoodCartScreen',
  PaymentFoodSuccess = 'PaymentFoodSuccessScreen',
  ReportReview = 'ReportReviewScreen',
  VideoDetail = 'VideoDetailScreen',
  ShopDetail = 'ShopDetailScreen',
  MyPoint = 'MyPointScreen',
  Notification = 'NotificationScreen',
  ShopFoodFavorite = 'ShopFoodFavoriteScreen',
  MyReview = 'MyReviewScreen',
  PersonalInformation = 'PersonalInformationScreen',
  ChangePassword = 'ChangePasswordScreen',
  Affiliate = 'AffiliateScreen',
  Login = 'LoginScreen',
  Term = 'TermScreen',
  OTP = 'OTPScreen',
  Register = 'RegisterScreen',
  Forgot = 'ForgotScreen',
  FoodOrderHistory = 'FoodOrderHistoryScreen',
  MyPage = 'MyPageScreen',
  Setting = 'SettingScreen',
  SearchResult = 'SearchResultScreen',
  FoodOrderDetail = 'FoodOrderDetailScreen',
  Content = 'ContentScreen',
  ResetPass = 'ResetPassScreen',
  Chat = 'ChatScreen',
  News = 'NewsScreen',
  NewsDetail = 'NewsDetailScreen',
  VerifyPhone = 'VerifyPhoneScreen',
  ReviewStore = 'ReviewStoreScreen'
}

export enum LocalStorageType {
  SpinWheel = "SPIN_WHEEL"
}